Si usano i double ora.

+++ Functions/ackley.h
@@ -11,9 +11,9 @@
 	#define MIN_SEARCH_RANGE -32768.0
 	#define MAX_SEARCH_RANGE 32768.0
 
-	float formulae(float values[]);
-		float firstCalculation(float values[]);
-		float secondCalculation(float values[]);
+	double formulae(double values[]);
+		double firstCalculation(double values[]);
+		double secondCalculation(double values[]);
 
 #endif
 #endif
